add lib/processing/graphics_context.rb
