+++ lib/processing.rb
@@ -15,4 +15,5 @@ require 'processing/font'
 require 'processing/touch'
 require 'processing/shader'
 require 'processing/capture'
+require 'processing/graphics_context'
 require 'processing/processing'
